+++ app/Http/routes.php
@@ -47,7 +47,7 @@ if (true === config('dfe.enable-console-api', false)) {
         \Route::get('instance/{instanceId}/services', 'Resources\LimitController@getInstanceServices');
         \Route::get('instance/{instanceId}/users', 'Resources\LimitController@getInstanceUsers');
         \Route::get('instance/{instanceId}/admins', 'Resources\LimitController@getInstanceAdmins');
-        \Route::post('instance/{instanceId}/delete', 'Resources\InstanceController@delete');
+        \Route::delete('instance/{instanceId}/delete', 'Resources\InstanceController@delete');
         \Route::get('reports/kibana', 'Resources\ReportController@getKibana');
 
         //  UI resource controllers